WIP: refactor of rich-text editor utility functions + tests
